+++ CHANGELOG.md
@@ -7,16 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [6.1.0] - Unreleased
 ### Added
-  - ?
+  - Console colored output (#50)
 ### Changed
   - Update Kotlin to `1.2.71` version
   - Update Gradle to `4.10.2` version
-  - Update default KtLint version to `0.27.0`
+  - Update default KtLint version to `0.29.0`
   - Hide specific source sets tasks
 ### Removed
   - ?
 ### Fixed
-  - ?
+  - Fixed plugin fails to configure android project with flavors (#131)
 
 ## [6.0.0] - 2018-9-20
 ### Added